+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_mpm.py
@@ -55,8 +55,17 @@ class TestMPM(unittest.TestCase):
         with self.assertRaises(NotImplementedError):
             pybamm.lithium_ion.MPM(options)
 
+    def test_differential_surface_form(self):
+        options = {"surface form": "differential"}
+        model = pybamm.lithium_ion.MPM(options)
+        model.check_well_posedness()
+
     def test_necessary_options(self):
         options = {"particle size": "single"}
+        with self.assertRaises(pybamm.OptionError):
+            pybamm.lithium_ion.MPM(options)
+
+        options = {"surface form": "false"}
         with self.assertRaises(pybamm.OptionError):
             pybamm.lithium_ion.MPM(options)